requirements.txt
Pillow>=9.0.0
numpy>=1.21.0
scikit-image>=0.19.0
imagehash>=4.3.0 